+++ src/__tests__/utils-test.ts
@@ -121,7 +121,6 @@ test('update navigation bar, with options', cb => {
     const options = {
         title: ANY_STRING,
         expandedTitle: ANY_STRING,
-        showExpandedTitle: false,
         showBackButton: true,
         showReloadButton: true,
         showProfileButton: false,
@@ -161,7 +160,6 @@ test('update navigation bar, without bridge', cb => {
     const options = {
         title: ANY_STRING,
         expandedTitle: ANY_STRING,
-        showExpandedTitle: false,
         showBackButton: true,
         showReloadButton: true,
         showProfileButton: false,