+++ src/utils.ts
@@ -41,7 +41,6 @@ export const share = (options: ShareOptions): Promise<void> =>
 export const updateNavigationBar = ({
     title,
     expandedTitle,
-    showExpandedTitle,
     showBackButton,
     showReloadButton,
     showProfileButton,
@@ -49,7 +48,6 @@ export const updateNavigationBar = ({
 }: {
     title?: string;
     expandedTitle?: string;
-    showExpandedTitle?: boolean;
     showBackButton?: boolean;
     showReloadButton?: boolean;
     showProfileButton?: boolean;
@@ -61,7 +59,6 @@ export const updateNavigationBar = ({
             payload: {
                 title,
                 expandedTitle,
-                showExpandedTitle,
                 showBackButton,
                 showReloadButton,
                 showProfileButton,